PARCIAL 2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
mariadb==1.1.7
bcrypt==4.0.1
uuid==1.30
pyyaml==6.0.1
mysql-connector-python==8.1.0